refactor: 肥大化したsimpleEditorを整理　#25
